Added multigridster and isValidMove
The multigridster allows you to move a gridster item from one gridster
to another. The only thing you have to do is pass the multiGridster:true
as a property to the gridster.

When a gridster item is moved from one gridster to another the functions
add and remove will be called so you can move the item from one array to
another. See the demo for more detailed example.

isValidMove allows you limit the possibilities of what the
height/width/col/row of a gridster item should be. Again see the demo
for more detailed example.